# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/dokter/eresep": {
            "get": {
                "tags": [
                    "Eresep"
                ],
                "summary": "Ambil semua data eresep",
                "operationId": "97a52a735aa6561cf7416a5e95e5bdb4",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data"
                    }
                }
            },
            "post": {
                "tags": [
                    "Eresep"
                ],
                "summary": "Simpan data eresep",
                "operationId": "4f71b980364f3265816defbd5a73a3c1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ID_RESEP",
                                    "ID_PEMERIKSAAN"
                                ],
                                "properties": {
                                    "ID_RESEP": {
                                        "type": "string",
                                        "example": "R001"
                                    },
                                    "ID_PEMERIKSAAN": {
                                        "type": "string",
                                        "example": "P001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data berhasil disimpan"
                    },
                    "400": {
                        "description": "Data tidak valid"
                    }
                }
            }
        },
        "/api/dokter/eresep/{id}": {
            "get": {
                "tags": [
                    "Eresep"
                ],
                "summary": "Ambil data eresep berdasarkan ID",
                "operationId": "70842caf38232abaef5dc089c867c70b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Eresep"
                ],
                "summary": "Update data eresep",
                "operationId": "8c06fb32388e3634beeebc4a6b934e6d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ID_RESEP",
                                    "ID_PEMERIKSAAN"
                                ],
                                "properties": {
                                    "ID_RESEP": {
                                        "type": "string",
                                        "example": "R001"
                                    },
                                    "ID_PEMERIKSAAN": {
                                        "type": "string",
                                        "example": "P001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data berhasil diupdate"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Eresep"
                ],
                "summary": "Hapus data eresep",
                "operationId": "25414dd7fa1f1f16c380630c28eb4fbe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berhasil dihapus"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            }
        },
...
    },
...
}